// HMSADMIN/src/client/app/staffmaster/staffmaster.ts
export class staffmasterTS { 
    StaffId: string; StaffFName: string; StaffMName: string; StaffLName: string; StaffName: string; StaffDob: string;
     CategoryIdStaffCategoryStaff: string; StaffPhoneNos: string; StaffJoinDate: string; StaffRetirementDate: string;
      StaffRemovalDate: string; StaffAddress1: string; StaffAddress2: string; StaffCity: string;
       StaffPin: string; StaffPhoneno: string; StaffMobileno: string; StaffEmailid: string; StaffFax: string;
        StaffSex: string;
         StaffMaritalStatus: string; StaffBloodGroup: string;
          StaffPost: string; StaffQualification: string;
   StaffRef: string;
    HospitalIdStaffMaster: string;
     CreatedateStaffMaster: string;
      ModifiedDateStaffMaster: string;
       StatusStaffMaster: string; 
}